# src/swagger/swagger.json
{
	"openapi": "3.0.0",
	"components": {
...
		"schemas": {
...
			"ObjectId": {
				"type": "string",
				"description": "A class representation of the BSON ObjectId type."
			},
...
			"ErrorResponse": {
				"properties": {
					"timestamp": {
						"type": "string"
					},
					"error": {
						"properties": {
							"details": {},
							"code": {
								"type": "string"
							},
							"message": {
								"type": "string"
							}
						},
						"required": [
							"message"
						],
						"type": "object"
					},
					"message": {
						"type": "string"
					},
					"success": {
						"type": "boolean",
						"enum": [
							false
						],
						"nullable": false
					}
				},
				"required": [
					"timestamp",
					"error",
					"message",
					"success"
				],
				"type": "object"
			},
...
			"SuccessResponse_null_": {
				"properties": {
					"pagination": {
						"properties": {
							"pages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"pages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					},
					"timestamp": {
						"type": "string"
					},
					"data": {
						"type": "number",
						"enum": [
							null
						],
						"nullable": true
					},
					"message": {
						"type": "string"
					},
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					}
				},
				"required": [
					"timestamp",
					"success"
				],
				"type": "object"
			},
			"ApiResponse_null_": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/SuccessResponse_null_"
					},
					{
						"$ref": "#/components/schemas/ErrorResponse"
					}
				]
			},
...
			"SuccessResponse__otpSent-boolean__": {
				"properties": {
					"pagination": {
						"properties": {
							"pages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"pages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					},
					"timestamp": {
						"type": "string"
					},
					"data": {
						"properties": {
							"otpSent": {
								"type": "boolean"
							}
						},
						"required": [
							"otpSent"
						],
						"type": "object"
					},
					"message": {
						"type": "string"
					},
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					}
				},
				"required": [
					"timestamp",
					"success"
				],
				"type": "object"
			},
			"SendOtpBody": {
				"properties": {
					"email": {
						"type": "string"
					},
					"mobile_number": {
						"type": "string"
					},
					"origin": {
						"type": "string"
					}
				},
				"type": "object",
				"additionalProperties": true
			},
			"SuccessResponse_any_": {
				"properties": {
					"pagination": {
						"properties": {
							"pages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"pages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					},
					"timestamp": {
						"type": "string"
					},
					"data": {},
					"message": {
						"type": "string"
					},
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					}
				},
				"required": [
					"timestamp",
					"success"
				],
				"type": "object"
			},
			"SuccessResponse____": {
				"properties": {
					"pagination": {
						"properties": {
							"pages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"pages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					},
					"timestamp": {
						"type": "string"
					},
					"data": {
						"properties": {},
						"type": "object"
					},
					"message": {
						"type": "string"
					},
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					}
				},
				"required": [
					"timestamp",
					"success"
				],
				"type": "object"
			},
			"OtpCheckBody": {
				"properties": {
					"otp": {
						"type": "string"
					},
					"registration_code": {
						"type": "string"
					}
				},
				"required": [
					"otp"
				],
				"type": "object",
				"additionalProperties": true
			},
			"SuccessResponse_unknown_": {
				"properties": {
					"pagination": {
						"properties": {
							"pages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"pages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					},
					"timestamp": {
						"type": "string"
					},
					"data": {},
					"message": {
						"type": "string"
					},
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					}
				},
				"required": [
					"timestamp",
					"success"
				],
				"type": "object"
			},
			"OtpCheckMobileBody": {
				"properties": {
					"otp": {
						"type": "string"
					},
					"requestId": {}
				},
				"required": [
					"otp"
				],
				"type": "object",
				"additionalProperties": true
			},
			"NewsletterSubscriber": {
				"properties": {
					"_id": {
						"$ref": "#/components/schemas/ObjectId"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					},
					"email": {
						"type": "string"
					},
					"active": {
						"type": "boolean"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"additionalProperties": true
			},
			"SuccessResponse_NewsletterSubscriber_": {
				"properties": {
					"pagination": {
						"properties": {
							"pages": {
								"type": "number",
								"format": "double"
							},
							"total": {
								"type": "number",
								"format": "double"
							},
							"limit": {
								"type": "number",
								"format": "double"
							},
							"page": {
								"type": "number",
								"format": "double"
							}
						},
						"required": [
							"pages",
							"total",
							"limit",
							"page"
						],
						"type": "object"
					},
					"timestamp": {
						"type": "string"
					},
					"data": {
						"$ref": "#/components/schemas/NewsletterSubscriber"
					},
					"message": {
						"type": "string"
					},
					"success": {
						"type": "boolean",
						"enum": [
							true
						],
						"nullable": false
					}
				},
				"required": [
					"timestamp",
					"success"
				],
				"type": "object"
			},
			"ApiResponse_NewsletterSubscriber_": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/SuccessResponse_NewsletterSubscriber_"
					},
					{
						"$ref": "#/components/schemas/ErrorResponse"
					}
				]
			},
			"Pick_NewsletterSubscriber.email-or-active_": {
				"properties": {
					"email": {
						"type": "string"
					},
					"active": {
						"type": "boolean"
					}
				},
				"required": [
					"email"
				],
				"type": "object",
				"description": "From T, pick a set of properties whose keys are in the union K"
			},
			"ApiResponse_any_": {
				"anyOf": [
					{
						"$ref": "#/components/schemas/SuccessResponse_any_"
					},
					{
						"$ref": "#/components/schemas/ErrorResponse"
					}
				]
			},
			"Partial_NewsletterSubscriber_": {
				"properties": {
					"email": {
						"type": "string"
					},
					"active": {
						"type": "boolean"
					},
					"_id": {
						"$ref": "#/components/schemas/ObjectId"
					},
					"createdAt": {
						"type": "string",
						"format": "date-time"
					},
					"updatedAt": {
						"type": "string",
						"format": "date-time"
					}
				},
				"type": "object",
				"description": "Make all properties in T optional"
			},
...
		},
		"securitySchemes": {}
	},
	"info": {
		"title": "app",
		"version": "1.0.0",
		"license": {
			"name": "ISC"
		},
		"contact": {}
	},
	"paths": {
...
		"/api/v1/otp/send-email": {
			"post": {
				"operationId": "SendOtpEmail",
				"responses": {
					"200": {
						"description": "OTP sent successfully",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ErrorResponse"
										},
										{
											"$ref": "#/components/schemas/SuccessResponse__otpSent-boolean__"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"otp"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SendOtpBody"
							}
						}
					}
				}
			}
		},
		"/api/v1/otp/send-mobile": {
			"post": {
				"operationId": "SendOtpMobile",
				"responses": {
					"200": {
						"description": "OTP sent successfully",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ErrorResponse"
										},
										{
											"$ref": "#/components/schemas/SuccessResponse_any_"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"otp"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/SendOtpBody"
							}
						}
					}
				}
			}
		},
		"/api/v1/otp/check-email": {
			"post": {
				"operationId": "CheckOtpEmail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ErrorResponse"
										},
										{
											"$ref": "#/components/schemas/SuccessResponse____"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"otp"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/OtpCheckBody"
							}
						}
					}
				}
			}
		},
		"/api/v1/otp/check-mobile": {
			"post": {
				"operationId": "CheckOtpMobile",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"anyOf": [
										{
											"$ref": "#/components/schemas/ErrorResponse"
										},
										{
											"$ref": "#/components/schemas/SuccessResponse_unknown_"
										}
									]
								}
							}
						}
					}
				},
				"tags": [
					"otp"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/OtpCheckMobileBody"
							}
						}
					}
				}
			}
		},
		"/api/v1/newsletter": {
			"post": {
				"operationId": "CreateSubscriber",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_NewsletterSubscriber_"
								}
							}
						}
					}
				},
				"tags": [
					"newsletter"
				],
				"security": [],
				"parameters": [],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Pick_NewsletterSubscriber.email-or-active_"
							}
						}
					}
				}
			},
			"get": {
				"operationId": "GetSubscribers",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_any_"
								}
							}
						}
					}
				},
				"tags": [
					"newsletter"
				],
				"security": [],
				"parameters": [
					{
						"in": "query",
						"name": "limit",
						"required": false,
						"schema": {
							"default": 10,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "skip",
						"required": false,
						"schema": {
							"default": 0,
							"format": "double",
							"type": "number"
						}
					},
					{
						"in": "query",
						"name": "sortBy",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"email",
								"createdAt",
								"_id",
								"updatedAt",
								"active"
							]
						}
					},
					{
						"in": "query",
						"name": "sortOrder",
						"required": false,
						"schema": {
							"type": "string",
							"enum": [
								"asc",
								"desc"
							]
						}
					},
					{
						"in": "query",
						"name": "filters",
						"required": false,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/newsletter/{id}": {
			"get": {
				"operationId": "GetSubscriber",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_NewsletterSubscriber_"
								}
							}
						}
					}
				},
				"tags": [
					"newsletter"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			},
			"put": {
				"operationId": "UpdateSubscriber",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_NewsletterSubscriber_"
								}
							}
						}
					}
				},
				"tags": [
					"newsletter"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				],
				"requestBody": {
					"required": true,
					"content": {
						"application/json": {
							"schema": {
								"$ref": "#/components/schemas/Partial_NewsletterSubscriber_"
							}
						}
					}
				}
			},
			"delete": {
				"operationId": "DeleteSubscriber",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_null_"
								}
							}
						}
					}
				},
				"tags": [
					"newsletter"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "id",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
		"/api/v1/newsletter/email/{email}": {
			"get": {
				"operationId": "GetSubscriberByEmail",
				"responses": {
					"200": {
						"description": "Ok",
						"content": {
							"application/json": {
								"schema": {
									"$ref": "#/components/schemas/ApiResponse_NewsletterSubscriber_"
								}
							}
						}
					}
				},
				"tags": [
					"newsletter"
				],
				"security": [],
				"parameters": [
					{
						"in": "path",
						"name": "email",
						"required": true,
						"schema": {
							"type": "string"
						}
					}
				]
			}
		},
...
	},
	"servers": [
		{
			"url": "/"
		}
	]
}